+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Modern JavaScript From The Beginning (Udemy)
   - Section 9: Error Handling & Regular Expressions
     - Regular Expressions [Part 4] - Shorthand Character Classes
+    - Regular Expressions - Form Validation Project
 - An example with shorthand charactee classes was added.
+- A small project with regular expressions was added.
 
 ## September 22, 2020
 - Modern JavaScript From The Beginning (Udemy)